reorganize directory & update imports
