+++ src/components/DataQuality/DataQuality.js
@@ -1,16 +1,16 @@
 /* eslint-disable react/jsx-filename-extension */
 import React from 'react';
 import { Switch, Route, useRouteMatch } from 'react-router-dom';
-import EfficiencyIntro from './EfficiencyIntro';
-import Stream from './Stream';
+import DataQualityIntro from './DataQualityIntro';
+import Stream from '../Stream';
 
-const Efficiency = function Efficiency() {
+const DataQuality = function DataQuality() {
   const match = useRouteMatch();
   return (
     <Stream>
       <Switch>
         <Route exact path={match.path}>
-          <EfficiencyIntro />
+          <DataQualityIntro />
         </Route>
       </Switch>
     </Stream>
@@ -18,4 +18,4 @@ const Efficiency = function Efficiency() {
   );
 };
 
-export default Efficiency;
+export default DataQuality;
